# Implement slowEqual enhancement from #128