This PR improves the speed of `slowEqual` by using [`bipf.createSeekPath`](https://github.com/ssbc/bipf/blob/master/index.js#L313) as suggested in #128.

## Before
  
  | Part | Speed | Heap Change | Samples |
  |---|---|---|---|
  | Query 1 big index with slowEqual (1st run) | 906.45ms ± 16.61ms | -851.7 kB ± 41.62 kB | 5 |
  | Query 1 big index with slowEqual (2nd run) | 245.64ms ± 7.76ms | -697.62 kB ± 38.45 kB | 5 |
  | Query 1 big index with slowEqual (3rd run) | 0.31ms ± 0.15ms | 38.13 kB ± 0 kB | 5 |


## After
  
  | Part | Speed | Heap Change | Samples |
  |---|---|---|---|
  | Query 1 big index with slowEqual (1st run) | 866.11ms ± 35ms | 1.07 MB ± 0 MB | 5 |
  | Query 1 big index with slowEqual (2nd run) | 296.57ms ± 4.96ms | 528.21 kB ± 0 kB | 5 |
  | Query 1 big index with slowEqual (3rd run) | 0.27ms ± 0.02ms | -85.7 kB ± 0 kB | 5 |