@arj03 2021-04-11
The second run seems slower?