@Barbarrosa 2021-04-11
@arj03 I ran this with the parallel benchmark, and those results also don't support using `bipf.createSeekPath`. I'll close this PR and note this on the original issue.

## [Before](https://github.com/Barbarrosa/jitdb/runs/2318803938?check_suite_focus=true)

  | Part | Speed | Heap Change | Samples |
  |---|---|---|---|
  | Query 1 big index with slowEqual (1st run) | 1227.72ms ± 30.83ms | -43.01 kB ± 18.86 kB | 45 |
  | Query 1 big index with slowEqual (2nd run) | 302.25ms ± 1.66ms | 77.83 kB ± 0 kB | 40 |
  | Query 1 big index with slowEqual (3rd run) | 0.29ms ± 0.02ms | 21.15 kB ± 0 kB | 44 |

## [After](https://github.com/Barbarrosa/jitdb/runs/2318797725?check_suite_focus=true)

  | Part | Speed | Heap Change | Samples |
  |---|---|---|---|
  | Query 1 big index with slowEqual (1st run) | 1407.86ms ± 54.24ms | 19.31 kB ± 0 kB | 39 |
  | Query 1 big index with slowEqual (2nd run) | 296.6ms ± 1.45ms | 1.95 kB ± 0 kB | 45 |
  | Query 1 big index with slowEqual (3rd run) | 0.36ms ± 0.03ms | 3 kB ± 26.53 kB | 48 |